Ported render code to typescript
* Created interface for Shaderbase
* Worked around heavy use of object literals and old school initialization
* TODO: Figure out why we also have pnpm and yarn. :/
* Added parcel config to enfoce TS validation at build time
