+++ .parcelrc
@@ -1,0 +1,8 @@
+{
+    "extends": "@parcel/config-default",
+    "validators": {
+        "./src/*.{ts,tsx}": [
+            "@parcel/validator-typescript"
+        ]
+    }
+}
