+++ .tool-versions
@@ -1,0 +1,1 @@
+nodejs 21.6.1
