+++ package.json
@@ -19,9 +19,11 @@
   "devDependencies": {
     "@parcel/reporter-bundle-analyzer": "^2.11.0",
     "@parcel/transformer-glsl": "^2.11.0",
+    "@parcel/validator-typescript": "^2.12.0",
     "@stylistic/eslint-plugin": "^1.6.0",
     "eslint": "^8.56.0",
-    "parcel": "^2.11.0"
+    "parcel": "^2.11.0",
+    "typescript": ">=3.0.0"
   },
   "scripts": {},
   "repository": {
