--- sculptgl.sublime-project
@@ -1,10 +1,0 @@
-{
-  "folders":[{
-      "name": "Sources",
-      "path": ".",
-    }, {
-      "name": "Resources",
-      "path": "app/resources",
-    }
-  ]
-}
