+++ tsconfig.json
@@ -9,5 +9,6 @@
         ],
         "noEmit": true,
         "noImplicitOverride": true,
+        "noImplicitThis": true,
     }
 }
